Make function range check more precise
Trying to be sure that try...catch clauses are handled correctly
I made the mistake to continue scanning intersection types for
any applicable intersections. This is imprecise. The scan should
stop when we 've covered the ArgTypes completely. I'm expecting
false positives on try..catch cases after this, but these are to
be handled by a better constraint generator.