Estimation of domain complexity
If we try to compare two syntactically different types for equality
we use t_elements on products and the result might be a large set
of simple domains. We try to prevent this by (over)estimating the
size of the elements list and collapsing the supremum if the
estimation is too big.