Better refinement of unexported functions' success typings
Unexported functions' success typings undergo a refinement process
in which the initial success typing is not (any, ..., any) -> any
but the supremum of all the calls towards this functions, which is
calculated during dataflow. This is of course a single clause
result which bundles everything in one constraint.

In this patch instead individual calls are kept separate and
create a disjunction. Disjunctive normal form splits it as usual
and in the end we catch errors like this one:

bar() -> a = foo(b).
buz() -> b = foo(a).

%% Unexported
foo(X) -> X.

... or gain clarity in situations like this:

foo(X) -> foo(X, []).

%% Unexported
foo(   [], A) -> A;
foo([H|T], A) -> foo(T, [H| A]).

In which former success typing was fun([_]) -> [_] while now it is
(([_,...]) -> [_,...]; ([]) -> []).