Prevent some extraneous range warnings
Specs may include a remote type in a range which is replaced with the
actual contents of that type. This had the side effect that whenever a
function would never return an element of that remote type we warned
about extraneous ranges. This is not desired as the usage of a remote
type is the correct way to describe the return of e.g. a wrapper to a
library function, even when the latter is called in a way that will
never return some of the possible results.