Unopaque contract and success typing range prior to 'extra' check
With the previous patch the extra range warnings are based on a
special solution of the remote types in a spec, excluding foreign
types, so as to not be warned if some of their elements are impossible
to be returned by the local function. This can trigger a wrong result
by t_inf for opaques, as the use of a foreign type within an opaque
type will produce different results between the full and the special
solution, that are handled erroneously by inf (which doesn't inspect
deeper structure but requires opaques to be identical).